modificacion de tipos de variables

+++ src/main/java/Mintictest/ciclo3/api/usuariosAPI.java
@@ -34,7 +34,7 @@ public class usuariosAPI {
 	}
 	
 	@DeleteMapping ("/eliminar/{id}")
-	public void eliminar (@PathVariable ("id") Integer id) {
+	public void eliminar (@PathVariable ("id") Long id) {
 		usuariosDAO.deleteById(id);
 	}
