+++ src/main/java/Mintictest/ciclo3/dao/clientesDAO.java
@@ -3,6 +3,6 @@ package Mintictest.ciclo3.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import Mintictest.ciclo3.model.clientes;
 
-public interface clientesDAO extends JpaRepository<clientes, Integer>{
+public interface clientesDAO extends JpaRepository<clientes, Long>{
 
 }
