+++ src/main/java/Mintictest/ciclo3/dao/usuariosDAO.java
@@ -4,6 +4,6 @@ package Mintictest.ciclo3.dao;
 import Mintictest.ciclo3.model.usuarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface usuariosDAO extends JpaRepository<usuarios, Integer>{
+public interface usuariosDAO extends JpaRepository<usuarios, Long>{
 	
 }
